requirements.txt
# # local packages
-e .
# external packages
albumentations
dvc
dvc-gdrive
ipykernel
jupyter
matplotlib
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
pyyaml
setuptools
scipy
scikit-learn
split-folders
torchsummary
ypstruct